+++ server/e2e/gql_me_test.go
@@ -1,20 +1,11 @@
 package e2e
 
 import (
-	"net/http"
 	"testing"
-
-	"github.com/reearth/reearth/server/internal/app/config"
 )
 
 func TestMe(t *testing.T) {
-	e := StartServer(t, &config.Config{
-		Origins: []string{"https://example.com"},
-		AuthSrv: config.AuthSrvConfig{
-			Disabled: true,
-		},
-	},
-		true, baseSeeder)
+	e := Server(t)
 
 	requestBody := GraphQLRequest{
 		OperationName: "GetMe",
@@ -22,15 +13,7 @@ func TestMe(t *testing.T) {
 		Variables:     map[string]any{},
 	}
 
-	e.POST("/api/graphql").
-		WithHeader("Origin", "https://example.com").
-		// WithHeader("authorization", "Bearer test").
-		WithHeader("X-Reearth-Debug-User", uID.String()).
-		WithHeader("Content-Type", "application/json").
-		WithJSON(requestBody).
-		Expect().
-		Status(http.StatusOK).
-		JSON().
+	Request(e, uID.String(), requestBody).
 		Object().
 		Value("data").Object().
 		Value("me").Object().
