+++ server/e2e/gql_scene_test.go
@@ -1,33 +1,14 @@
 package e2e
 
 import (
-	"context"
 	"testing"
 
-	"github.com/reearth/reearth/server/internal/app/config"
-	"github.com/reearth/reearth/server/internal/usecase/repo"
 	"golang.org/x/text/language"
 )
 
-// export REEARTH_DB=mongodb://localhost
-
-// go test -v -run TestGetScenePlaceholderEnglish ./e2e/...
 func TestGetScenePlaceholderEnglish(t *testing.T) {
 
-	e := StartServer(
-		t,
-		&config.Config{
-			Origins: []string{"https://example.com"},
-			AuthSrv: config.AuthSrvConfig{
-				Disabled: true,
-			},
-		},
-		true,
-		// English user Seeder
-		func(ctx context.Context, r *repo.Container) error {
-			return baseSeederWithLang(ctx, r, language.English)
-		},
-	)
+	e := ServerEnglish(t)
 	pID := createProjectWithExternalImage(e, "test")
 	_, _, sID := createScene(e, pID)
 	r := getScene(e, sID, language.English.String())
@@ -53,24 +34,9 @@ func TestGetScenePlaceholderEnglish(t *testing.T) {
 
 }
 
-// go test -v -run TestGetScenePlaceholderJapanese ./e2e/...
-
 func TestGetScenePlaceholderJapanese(t *testing.T) {
 
-	e := StartServer(
-		t,
-		&config.Config{
-			Origins: []string{"https://example.com"},
-			AuthSrv: config.AuthSrvConfig{
-				Disabled: true,
-			},
-		},
-		true,
-		// Japanese user Seeder
-		func(ctx context.Context, r *repo.Container) error {
-			return baseSeederWithLang(ctx, r, language.Japanese)
-		},
-	)
+	e := ServerJapanese(t)
 	pID := createProjectWithExternalImage(e, "test")
 	_, _, sID := createScene(e, pID)
 	r := getScene(e, sID, language.Japanese.String())
