+++ server/e2e/gql_style_test.go
@@ -1,11 +1,9 @@
 package e2e
 
 import (
-	"net/http"
 	"testing"
 
 	"github.com/gavv/httpexpect/v2"
-	"github.com/reearth/reearth/server/internal/app/config"
 )
 
 func addStyle(e *httpexpect.Expect, sId, name string) (GraphQLRequest, *httpexpect.Value, string) {
@@ -51,12 +49,7 @@ func addStyle(e *httpexpect.Expect, sId, name string) (GraphQLRequest, *httpexpe
 		},
 	}
 
-	res := e.POST("/api/graphql").
-		WithHeader("Content-Type", "application/json").
-		WithJSON(requestBody).
-		Expect().
-		Status(http.StatusOK).
-		JSON()
+	res := Request(e, uID.String(), requestBody)
 
 	styleId := res.Path("$.data.addStyle.style.id").String().Raw()
 	return requestBody, res, styleId
@@ -80,12 +73,7 @@ func updateStyleName(e *httpexpect.Expect, styleId, newName string) (GraphQLRequ
 		},
 	}
 
-	res := e.POST("/api/graphql").
-		WithHeader("Content-Type", "application/json").
-		WithJSON(requestBody).
-		Expect().
-		Status(http.StatusOK).
-		JSON()
+	res := Request(e, uID.String(), requestBody)
 
 	return requestBody, res
 }
@@ -103,12 +91,7 @@ func removeStyle(e *httpexpect.Expect, styleId string) (GraphQLRequest, *httpexp
 		},
 	}
 
-	res := e.POST("/api/graphql").
-		WithHeader("Content-Type", "application/json").
-		WithJSON(requestBody).
-		Expect().
-		Status(http.StatusOK).
-		JSON()
+	res := Request(e, uID.String(), requestBody)
 
 	return requestBody, res
 }
@@ -130,12 +113,7 @@ func duplicateStyle(e *httpexpect.Expect, styleId string) (GraphQLRequest, *http
 		},
 	}
 
-	res := e.POST("/api/graphql").
-		WithHeader("Content-Type", "application/json").
-		WithJSON(requestBody).
-		Expect().
-		Status(http.StatusOK).
-		JSON()
+	res := Request(e, uID.String(), requestBody)
 
 	return requestBody, res
 }
@@ -162,25 +140,13 @@ func fetchSceneForStyles(e *httpexpect.Expect, sID string) (GraphQLRequest, *htt
 		},
 	}
 
-	res := e.POST("/api/graphql").
-		WithHeader("Origin", "https://example.com").
-		WithHeader("X-Reearth-Debug-User", uID.String()).
-		WithHeader("Content-Type", "application/json").
-		WithJSON(fetchSceneRequestBody).
-		Expect().
-		Status(http.StatusOK).
-		JSON()
+	res := Request(e, uID.String(), fetchSceneRequestBody)
 
 	return fetchSceneRequestBody, res
 }
 
 func TestStyleCRUD(t *testing.T) {
-	e := StartServer(t, &config.Config{
-		Origins: []string{"https://example.com"},
-		AuthSrv: config.AuthSrvConfig{
-			Disabled: true,
-		},
-	}, true, baseSeeder)
+	e := Server(t)
 
 	pId := createProject(e, "test")
 	_, _, sId := createScene(e, pId)
