+++ server/e2e/seeder.go
@@ -12,6 +12,7 @@ import (
 	"github.com/reearth/reearthx/account/accountdomain"
 	"github.com/reearth/reearthx/account/accountdomain/user"
 	"github.com/reearth/reearthx/account/accountdomain/workspace"
+	"github.com/reearth/reearthx/idx"
 	"github.com/reearth/reearthx/util"
 	"github.com/samber/lo"
 	"golang.org/x/text/language"
@@ -35,6 +36,13 @@ var (
 	dsfID4 = dataset.NewFieldID()
 
 	now = time.Date(2022, time.January, 1, 0, 0, 0, 0, time.UTC)
+
+	uId1 = accountdomain.NewUserID()
+	uId2 = accountdomain.NewUserID()
+	uId3 = accountdomain.NewUserID()
+	wId1 = accountdomain.NewWorkspaceID()
+	wId2 = accountdomain.NewWorkspaceID()
+	iId1 = accountdomain.NewIntegrationID()
 )
 
 func baseSeeder(ctx context.Context, r *repo.Container) error {
@@ -117,6 +125,74 @@ func baseSetup(ctx context.Context, r *repo.Container, u *user.User) error {
 		}).
 		MustBuild()
 	if err := r.Dataset.Save(ctx, ds); err != nil {
+		return err
+	}
+
+	return nil
+}
+
+func baseSeederUser(ctx context.Context, r *repo.Container) error {
+	auth := user.ReearthSub(uId1.String())
+	u := user.New().ID(uId1).
+		Name("e2e").
+		Email("e2e@e2e.com").
+		Auths([]user.Auth{*auth}).
+		Theme(user.ThemeDark).
+		Lang(language.Japanese).
+		Workspace(wId1).
+		MustBuild()
+	if err := r.User.Save(ctx, u); err != nil {
+		return err
+	}
+	u2 := user.New().ID(uId2).
+		Name("e2e2").
+		Workspace(wId2).
+		Email("e2e2@e2e.com").
+		MustBuild()
+	if err := r.User.Save(ctx, u2); err != nil {
+		return err
+	}
+	u3 := user.New().ID(uId3).
+		Name("e2e3").
+		Workspace(wId2).
+		Email("e2e3@e2e.com").
+		MustBuild()
+	if err := r.User.Save(ctx, u3); err != nil {
+		return err
+	}
+	roleOwner := workspace.Member{
+		Role:      workspace.RoleOwner,
+		InvitedBy: uId1,
+	}
+	roleReader := workspace.Member{
+		Role:      workspace.RoleReader,
+		InvitedBy: uId2,
+	}
+
+	w := workspace.New().ID(wId1).
+		Name("e2e").
+		Members(map[idx.ID[accountdomain.User]]workspace.Member{
+			uId1: roleOwner,
+		}).
+		Integrations(map[idx.ID[accountdomain.Integration]]workspace.Member{
+			iId1: roleOwner,
+		}).
+		MustBuild()
+	if err := r.Workspace.Save(ctx, w); err != nil {
+		return err
+	}
+
+	w2 := workspace.New().ID(wId2).
+		Name("e2e2").
+		Members(map[idx.ID[accountdomain.User]]workspace.Member{
+			uId1: roleOwner,
+			uId3: roleReader,
+		}).
+		Integrations(map[idx.ID[accountdomain.Integration]]workspace.Member{
+			iId1: roleOwner,
+		}).
+		MustBuild()
+	if err := r.Workspace.Save(ctx, w2); err != nil {
 		return err
 	}
